+++ src/commands/command.rs
@@ -1,6 +1,10 @@
 use crate::{
     Context, HandlerResult,
-    commands::{set_currency, set_language, show_balance, show_balances, show_expense},
+    commands::{
+        HelpMessage, add_traveler, delete_expense, delete_transfer, delete_traveler, help,
+        list_expenses, list_transfers, list_travelers, set_currency, set_language, show_balances,
+        show_expense, transfer,
+    },
     i18n::{self, Translatable, help::*, translate, translate_with_args},
     traveler::Name,
 };
@@ -11,11 +15,6 @@ use std::sync::{Arc, Mutex};
 use strum::{AsRefStr, EnumIter, EnumString, IntoEnumIterator};
 use teloxide::{prelude::*, utils::command::BotCommands};
 use unic_langid::LanguageIdentifier;
-
-use crate::commands::{
-    HelpMessage, add_traveler, delete_expense, delete_traveler, find_expenses, help, list_expenses,
-    list_travelers, transfer,
-};
 
 pub static COMMANDS: LazyLock<Vec<String>> = LazyLock::new(|| {
     Command::iter()
@@ -44,9 +43,7 @@ pub enum Command {
     #[command(description = "{descr-delete-expense}")]
     DeleteExpense { number: i64 },
     #[command(description = "{descr-list-expenses}")]
-    ListExpenses,
-    #[command(description = "{descr-find-expenses}")]
-    FindExpenses { description: String },
+    ListExpenses { description: String },
     #[command(description = "{descr-show-expense}")]
     ShowExpense { number: i64 },
     #[command(description = "{descr-transfer}", parse_with = "split")]
@@ -55,10 +52,12 @@ pub enum Command {
         to: Name,
         amount: Decimal,
     },
-    #[command(description = "{descr-show-balance}")]
-    ShowBalance { name: Name },
+    #[command(description = "{descr-delete-transfer}")]
+    DeleteTransfer { number: i64 },
+    #[command(description = "{descr-list-transfers}")]
+    ListTransfers { name: Name },
     #[command(description = "{descr-show-balances}")]
-    ShowBalances,
+    ShowBalances { name: Name },
     #[command(description = "{descr-cancel}")]
     Cancel,
 }
@@ -94,16 +93,16 @@ impl HelpMessage for Command {
             ListTravelers => translate(ctx, HELP_LIST_TRAVELERS),
             AddExpense => translate(ctx, HELP_ADD_EXPENSE),
             DeleteExpense { number: _ } => translate(ctx, HELP_DELETE_EXPENSE),
-            ListExpenses => translate(ctx, HELP_LIST_EXPENSES),
-            FindExpenses { description: _ } => translate(ctx, HELP_FIND_EXPENSES),
+            ListExpenses { description: _ } => translate(ctx, HELP_LIST_EXPENSES),
             ShowExpense { number: _ } => translate(ctx, HELP_SHOW_EXPENSE),
             Transfer {
                 from: _,
                 to: _,
                 amount: _,
             } => translate(ctx, HELP_TRANSFER),
-            ShowBalance { name: _ } => translate(ctx, HELP_SHOW_BALANCE),
-            ShowBalances => translate(ctx, HELP_SHOW_BALANCES),
+            DeleteTransfer { number: _ } => translate(ctx, HELP_DELETE_TRANSFER),
+            ListTransfers { name: _ } => translate(ctx, HELP_LIST_TRANSFERS),
+            ShowBalances { name: _ } => translate(ctx, HELP_SHOW_BALANCES),
             Cancel => translate(ctx, HELP_CANCEL),
         }
     }
@@ -125,12 +124,12 @@ pub async fn commands_handler(
         DeleteTraveler { name } => delete_traveler(&msg, name, ctx.clone()).await,
         ListTravelers => list_travelers(&msg, ctx.clone()).await,
         DeleteExpense { number } => delete_expense(&msg, number, ctx.clone()).await,
-        ListExpenses => list_expenses(&msg, ctx.clone()).await,
-        FindExpenses { description } => find_expenses(&msg, &description, ctx.clone()).await,
+        ListExpenses { description } => list_expenses(&msg, &description, ctx.clone()).await,
         ShowExpense { number } => show_expense(&msg, number, ctx.clone()).await,
         Transfer { from, to, amount } => transfer(&msg, from, to, amount, ctx.clone()).await,
-        ShowBalance { name } => show_balance(&msg, name, ctx.clone()).await,
-        ShowBalances => show_balances(&msg, ctx.clone()).await,
+        DeleteTransfer { number } => delete_transfer(&msg, number, ctx.clone()).await,
+        ListTransfers { name } => list_transfers(&msg, name, ctx.clone()).await,
+        ShowBalances { name } => show_balances(&msg, name, ctx.clone()).await,
         Cancel | AddExpense => {
             unreachable!("This command is handled before calling this function.")
         }